ml-job-worker/requirements.txt
absl-py==2.2.2 ; python_version == "3.11"
amqp==5.3.1 ; python_version == "3.11"
annotated-types==0.7.0 ; python_version == "3.11"
anyio==4.9.0 ; python_version == "3.11"
astropy-iers-data==0.2025.4.28.0.37.27 ; python_version == "3.11"
astropy==7.0.1 ; python_version == "3.11"
astunparse==1.6.3 ; python_version == "3.11"
billiard==4.2.1 ; python_version == "3.11"
black==25.1.0 ; python_version == "3.11"
celery==5.5.2 ; python_version == "3.11"
certifi==2025.4.26 ; python_version == "3.11"
charset-normalizer==3.4.1 ; python_version == "3.11"
click-didyoumean==0.3.1 ; python_version == "3.11"
click-plugins==1.1.1 ; python_version == "3.11"
click-repl==0.3.0 ; python_version == "3.11"
click==8.1.8 ; python_version == "3.11"
colorama==0.4.6 ; python_version == "3.11" and platform_system == "Windows"
contourpy==1.3.2 ; python_version == "3.11"
cycler==0.12.1 ; python_version == "3.11"
flatbuffers==25.2.10 ; python_version == "3.11"
fonttools==4.57.0 ; python_version == "3.11"
gast==0.6.0 ; python_version == "3.11"
google-pasta==0.2.0 ; python_version == "3.11"
grpcio==1.71.0 ; python_version == "3.11"
h11==0.16.0 ; python_version == "3.11"
h5py==3.13.0 ; python_version == "3.11"
httpcore==1.0.9 ; python_version == "3.11"
httpx==0.28.1 ; python_version == "3.11"
idna==3.10 ; python_version == "3.11"
imbalanced-learn==0.13.0 ; python_version == "3.11"
imblearn==0.0 ; python_version == "3.11"
isort==6.0.1 ; python_version == "3.11"
joblib==1.4.2 ; python_version == "3.11"
keras==3.9.2 ; python_version == "3.11"
kiwisolver==1.4.8 ; python_version == "3.11"
kombu==5.5.3 ; python_version == "3.11"
libclang==18.1.1 ; python_version == "3.11"
markdown-it-py==3.0.0 ; python_version == "3.11"
markdown==3.8 ; python_version == "3.11"
markupsafe==3.0.2 ; python_version == "3.11"
matplotlib==3.10.1 ; python_version == "3.11"
mdurl==0.1.2 ; python_version == "3.11"
ml-dtypes==0.3.2 ; python_version == "3.11"
mypy-extensions==1.1.0 ; python_version == "3.11"
namex==0.0.9 ; python_version == "3.11"
numpy==1.26.4 ; python_version == "3.11"
nvidia-cublas-cu12==12.3.4.1 ; python_version == "3.11"
nvidia-cuda-cupti-cu12==12.3.101 ; python_version == "3.11"
nvidia-cuda-nvcc-cu12==12.3.107 ; python_version == "3.11"
nvidia-cuda-nvrtc-cu12==12.3.107 ; python_version == "3.11"
nvidia-cuda-runtime-cu12==12.3.101 ; python_version == "3.11"
nvidia-cudnn-cu12==8.9.7.29 ; python_version == "3.11"
nvidia-cufft-cu12==11.0.12.1 ; python_version == "3.11"
nvidia-curand-cu12==10.3.4.107 ; python_version == "3.11"
nvidia-cusolver-cu12==11.5.4.101 ; python_version == "3.11"
nvidia-cusparse-cu12==12.2.0.103 ; python_version == "3.11"
nvidia-nccl-cu12==2.19.3 ; python_version == "3.11"
nvidia-nvjitlink-cu12==12.3.101 ; python_version == "3.11"
opt-einsum==3.4.0 ; python_version == "3.11"
optree==0.15.0 ; python_version == "3.11"
packaging==25.0 ; python_version == "3.11"
pathspec==0.12.1 ; python_version == "3.11"
pillow==11.2.1 ; python_version == "3.11"
platformdirs==4.3.7 ; python_version == "3.11"
prompt-toolkit==3.0.51 ; python_version == "3.11"
protobuf==4.25.7 ; python_version == "3.11"
pydantic-core==2.33.2 ; python_version == "3.11"
pydantic-settings==2.9.1 ; python_version == "3.11"
pydantic==2.11.4 ; python_version == "3.11"
pyerfa==2.0.1.5 ; python_version == "3.11"
pygments==2.19.1 ; python_version == "3.11"
pyparsing==3.2.3 ; python_version == "3.11"
python-dateutil==2.9.0.post0 ; python_version == "3.11"
python-dotenv==1.1.0 ; python_version == "3.11"
pyyaml==6.0.2 ; python_version == "3.11"
requests==2.32.3 ; python_version == "3.11"
rich==14.0.0 ; python_version == "3.11"
scikit-learn==1.6.1 ; python_version == "3.11"
scipy==1.15.2 ; python_version == "3.11"
#setuptools==80.1.0 ; python_version == "3.11"
six==1.17.0 ; python_version == "3.11"
sklearn-compat==0.1.3 ; python_version == "3.11"
sniffio==1.3.1 ; python_version == "3.11"
tensorboard-data-server==0.7.2 ; python_version == "3.11"
tensorboard==2.16.2 ; python_version == "3.11"
tensorflow-io-gcs-filesystem==0.37.1 ; python_version == "3.11"
tensorflow==2.16.2 ; python_version == "3.11"
termcolor==3.1.0 ; python_version == "3.11"
threadpoolctl==3.6.0 ; python_version == "3.11"
typing-extensions==4.13.2 ; python_version == "3.11"
typing-inspection==0.4.0 ; python_version == "3.11"
tzdata==2025.2 ; python_version == "3.11"
urllib3==2.4.0 ; python_version == "3.11"
vine==5.1.0 ; python_version == "3.11"
wcwidth==0.2.13 ; python_version == "3.11"
werkzeug==3.1.3 ; python_version == "3.11"
wheel==0.45.1 ; python_version == "3.11"
wrapt==1.17.2 ; python_version == "3.11"
